Extract the comparers
The individual comparisons have been extracted into
separate classes.
