+++ src/NUnitFramework/framework/Constraints/Comparers/DictionariesComparer.cs
@@ -1,0 +1,56 @@
+﻿// ***********************************************************************
+// Copyright (c) 2009 Charlie Poole, Rob Prouse
+//
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject to
+// the following conditions:
+//
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+// ***********************************************************************
+
+using System.Collections;
+
+namespace NUnit.Framework.Constraints.Comparers
+{
+    /// <summary>
+    /// Comparator for two <see cref="IDictionary"/>s.
+    /// </summary>
+    internal class DictionariesComparer
+    {
+        private readonly NUnitEqualityComparer _equalityComparer;
+
+        internal DictionariesComparer(NUnitEqualityComparer equalityComparer)
+        {
+            _equalityComparer = equalityComparer;
+        }
+
+        internal bool Equal(IDictionary x, IDictionary y, ref Tolerance tolerance)
+        {
+            if (x.Count != y.Count)
+                return false;
+
+            CollectionTally tally = new CollectionTally(_equalityComparer, x.Keys);
+            if (!tally.TryRemove(y.Keys) || tally.Count > 0)
+                return false;
+
+            foreach (object key in x.Keys)
+                if (!_equalityComparer.AreEqual(x[key], y[key], ref tolerance))
+                    return false;
+
+            return true;
+        }
+    }
+}
