+++ src/NUnitFramework/framework/Constraints/EqualConstraint.cs
@@ -46,7 +46,7 @@ namespace NUnit.Framework.Constraints
         /// <summary>
         /// NUnitEqualityComparer used to test equality.
         /// </summary>
-        private NUnitEqualityComparer _comparer = new NUnitEqualityComparer();
+        private NUnitEqualityComparer _comparer = NUnitEqualityComparer.Default;
 
         #endregion
