+++ src/NUnitFramework/framework/nunit.framework-2.0.csproj
@@ -88,6 +88,19 @@
     <Compile Include="Compatibility\System.Threading\SpinWait.cs" />
     <Compile Include="Compatibility\System\Lazy.cs" />
     <Compile Include="Constraints\CollectionSupersetConstraint.cs" />
+    <Compile Include="Constraints\Comparers\ArraysComparer.cs" />
+    <Compile Include="Constraints\Comparers\CharsComparer.cs" />
+    <Compile Include="Constraints\Comparers\DateTimeOffsetsComparer.cs" />
+    <Compile Include="Constraints\Comparers\DictionariesComparer.cs" />
+    <Compile Include="Constraints\Comparers\DictionaryEntriesComparer.cs" />
+    <Compile Include="Constraints\Comparers\DirectoriesComparer.cs" />
+    <Compile Include="Constraints\Comparers\EnumerablesComparer.cs" />
+    <Compile Include="Constraints\Comparers\IEquatablesComparer.cs" />
+    <Compile Include="Constraints\Comparers\KeyValuePairsComparer.cs" />
+    <Compile Include="Constraints\Comparers\NumericsComparer.cs" />
+    <Compile Include="Constraints\Comparers\StreamsComparer.cs" />
+    <Compile Include="Constraints\Comparers\StringsComparer.cs" />
+    <Compile Include="Constraints\Comparers\TimeSpanToleranceComparer.cs" />
     <Compile Include="Constraints\DictionaryContainsValueConstraint.cs" />
     <Compile Include="Constraints\EqualConstraintResult.cs" />
     <Compile Include="Constraints\ExceptionNotThrownConstraint.cs" />
