+++ src/NUnitFramework/framework/nunit.framework-netstandard13.csproj
@@ -166,6 +166,19 @@
     <Compile Include="Constraints\CollectionSubsetConstraint.cs" />
     <Compile Include="Constraints\CollectionSupersetConstraint.cs" />
     <Compile Include="Constraints\CollectionTally.cs" />
+    <Compile Include="Constraints\Comparers\ArraysComparer.cs" />
+    <Compile Include="Constraints\Comparers\CharsComparer.cs" />
+    <Compile Include="Constraints\Comparers\DateTimeOffsetsComparer.cs" />
+    <Compile Include="Constraints\Comparers\DictionariesComparer.cs" />
+    <Compile Include="Constraints\Comparers\DictionaryEntriesComparer.cs" />
+    <Compile Include="Constraints\Comparers\DirectoriesComparer.cs" />
+    <Compile Include="Constraints\Comparers\EnumerablesComparer.cs" />
+    <Compile Include="Constraints\Comparers\IEquatablesComparer.cs" />
+    <Compile Include="Constraints\Comparers\KeyValuePairsComparer.cs" />
+    <Compile Include="Constraints\Comparers\NumericsComparer.cs" />
+    <Compile Include="Constraints\Comparers\StreamsComparer.cs" />
+    <Compile Include="Constraints\Comparers\StringsComparer.cs" />
+    <Compile Include="Constraints\Comparers\TimeSpanToleranceComparer.cs" />
     <Compile Include="Constraints\ComparisonAdapter.cs" />
     <Compile Include="Constraints\ComparisonConstraint.cs" />
     <Compile Include="Constraints\Constraint.cs" />
